+++ CHANGELOG.md
@@ -5,8 +5,7 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
-
-## **WORK IN PROGRESS**
+## 1.0.59 (2022-08-21)
 
 * (theimo1221)  Fix import resolve